requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.59
langchain-community>=0.0.29
langchain-openai>=0.0.8
langchain-google-genai>=0.0.4
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.66.0
pypdf>=3.17.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0